refactor(repodata-gateway): move the boxed future/stream helpers to a shared module
Addresses review feedback: who_needs_query no longer depends on types
declared in query.rs, and Gateway::who_needs gets a user facing doc
comment with an example instead of a fully qualified target type.
